// src/app/views/nsa/models/IsmsReportResponse.ts
export class IsmsReportResponse {
    requisitionNo: string ;
	Product: string ;
	msisdn: string ;
	SIM: string ;
	requisitionType: string ;
	requesterName: string ;
	requesterMobile: string;
	msisdnStatus: string ;
	startDate: string ;
	endDate: string ;
	creditLimit: string ;
	newCreditLimit: string ;
	rechargeAmount: string ;
	newOwner: string ;
	newEndDate: string ;
}